separating the camera transform math
